move to circuits.common
